chore(viewer): regenerate the registry consumer for this branch's source
This branch extracts the camera fit into its own file and rewires the viewer
around it, so the item gained a file and lost lines from another; the committed
`shadcn add` output has to be regenerated with them — the copy is what CI typechecks and
serves, which is how the registry is dogfooded.

As on the branch below, `sync:registry` also re-resolves `pnpm-lock.yaml` and pins
the example's dependency ranges to the installed versions. Both are reverted:
CI installs with `--frozen-lockfile`, and the ranges the example declares are
satisfied by what is installed. `pnpm install --frozen-lockfile` and the example's
`pnpm typecheck` pass against the regenerated copy.
